move stub method in stub helper
